move calendar UX changes to utils file
