+++ src/claimApplication/utils.js
@@ -17,6 +17,9 @@ export const STORAGE_KEY_EDITING_WORKERS_COMP = "editing_workers_comp";
 
 export const STORAGE_KEY_SESSION_DATA = "session_data";
 
+export const CALENDAR_CONTROL_ID = "CalendarControl";
+export const FDD_CALENDAR_CONTROL_ID = "FDDCalendarControl"; // DEPENDENT on FDD, not setting it
+
 export function getSessionData() {
   try {
     const encodedData = sessionStorage.getItem(STORAGE_KEY_SESSION_DATA);
@@ -264,4 +267,105 @@ export function removeIntroTextReferencingFuture() {
       el.innerHTML = el.innerHTML.replace(targetRegex, '');
     }
   });
+}
+
+export function updateCalendarUI(id, calendarId) {
+  const element = document.getElementById(id);
+  if (element) {
+    element.src = "https://beta.nj.gov/files/tdi-fli-claim-status/assets/calendar_today.svg";
+    element.addEventListener('click', function() {
+      closeCalendarPopup(calendarId);
+      setTimeout(() => fixCalendarPopup(calendarId), 10);
+    });
+  }
+}
+
+function fixCalendarPopup(calendarId) {
+  const headerRows = document.querySelectorAll(`#${calendarId} tr.header`);
+  if (headerRows.length != 2) return;
+
+  const navHeaderRow = headerRows[0];
+  const footerRow = headerRows[1];
+
+  const functionSuffix = calendarId === FDD_CALENDAR_CONTROL_ID ? 'FDD' : '';
+  const linkStyle = 'font-size: 1.2em; padding: 4px; display: inline-flex; align-items: center; justify-content: center; min-width: 22px; min-height: 22px; text-decoration: none; color: inherit;';
+  const imgStyle = 'width: 20px; height: 20px; display: block;';
+  
+  const baseUrl = 'https://beta.nj.gov/files/tdi-fli-claim-status/assets/';
+  const prevYear = `<a href="javascript:changeCalendarControlYear${functionSuffix}(-1);" style="${linkStyle}"><img src="${baseUrl}navigate_far_before.svg" alt="Previous year" style="${imgStyle}"></a>`;
+  const prevMonth = `<a href="javascript:changeCalendarControlMonth${functionSuffix}(-1);" style="${linkStyle}"><img src="${baseUrl}navigate_before.svg" alt="Previous month" style="${imgStyle}"></a>`;
+  const nextMonth = `<a href="javascript:changeCalendarControlMonth${functionSuffix}(1);" style="${linkStyle}"><img src="${baseUrl}navigate_next.svg" alt="Next month" style="${imgStyle}"></a>`;
+  const nextYear = `<a href="javascript:changeCalendarControlYear${functionSuffix}(1);" style="${linkStyle}"><img src="${baseUrl}navigate_far_next.svg" alt="Next year" style="${imgStyle}"></a>`;
+  
+  const title = navHeaderRow.querySelector('.title')?.innerHTML || '';
+  
+  const gapStyle = 'display: flex; gap: 12px; align-items: center;';
+  
+  navHeaderRow.innerHTML = `
+    <td colspan="7" style="padding: 0;">
+      <div style="display: flex; justify-content: space-between; align-items: center; padding: 8px 4px;">
+        <div style="text-align: left; ${gapStyle}">
+          ${prevYear}
+          ${prevMonth}
+        </div>
+        <div class="title" style="text-align: center; flex: 1;">${title}</div>
+        <div style="text-align: right; ${gapStyle}">
+          ${nextMonth}
+          ${nextYear}
+        </div>
+      </div>
+    </td>
+  `;
+
+  removeEmptyRows(calendarId);
+
+  const navLinks = navHeaderRow.querySelectorAll('a');
+  navLinks.forEach(link => {
+    link.addEventListener('click', function() {
+      setTimeout(() => fixCalendarPopup(calendarId), 10);
+    });
+  });
+
+  footerRow.remove();
+}
+
+function closeCalendarPopup(calendarId) {
+  const calendarPopup = document.getElementById(calendarId);
+  
+  if (window.calendarClickOutsideHandler) {
+    document.removeEventListener('click', window.calendarClickOutsideHandler);
+  }
+  
+  window.calendarClickOutsideHandler = function(event) {
+    if (!calendarPopup || !calendarPopup.offsetParent) return;
+    const isClickInside = calendarPopup.contains(event.target);
+    if (!isClickInside) {
+      if (calendarId === CALENDAR_CONTROL_ID) {
+        hideCalendarControl();
+      } else if (calendarId === FDD_CALENDAR_CONTROL_ID) {
+        hideCalendarControlFDD();
+      }
+      document.removeEventListener('click', window.calendarClickOutsideHandler);
+    }
+  };
+  
+  setTimeout(() => {
+    document.addEventListener('click', window.calendarClickOutsideHandler);
+  }, 100);
+}
+
+function removeEmptyRows(calendarId) {
+  const calendarTable = document.querySelector(`#${calendarId} table`);
+  if (calendarTable) {
+    const allRows = calendarTable.querySelectorAll('tr');
+    allRows.forEach(row => {
+      const cells = row.querySelectorAll('td');
+      if (cells.length > 0) {
+        const allEmpty = Array.from(cells).every(cell => cell.classList.contains('empty'));
+        if (allEmpty) {
+          row.remove();
+        }
+      }
+    });
+  }
 }
